A version of IBM Model 1 as described in Koehn 2010.
Ported from implementation in http://code.google.com/p/clir/.

Unfortunately the documentation is a little unclear as to how to handle NULL 
tokens in the algorithm. Most of the examples given ignore the NULL token 
which makes it slightly difficult to come up with some more comprehensive 
tests.


svn/trunk@8695