wordnet reader is fixed for pypy and py3 (and its operation is restored for py26 and py27).
Wordnet tests were always failing for pypy because it is not guaranteed that strings with the same content have the same identity. After switching WordNetCorpusReader to unicode tests began to fail for python 2.6 and 2.7 because unicode string is not the same as byte string constant (even if they are equal and == returns True). This one was quite tricky to debug :)