fix malt's mco parameter documentation
As far as I could see, MaltParser does not take a full path as input for
mco models. You must specify the name of the model, which MaltParser
then uses to look for a file in ${working_dir}/${mco}.mco. It's a bit
clunky, but I couldn't find alternatives in their documentation.