Substantial rewrite of the Porter Stemmer. This was to: 1) Make it threadsafe 2) Make it more Pythonic and remove at least some of the weird C-like code from    when it was originally ported from C 3) Remove nonsense comments referring to things like null terminated strings and    char arrays that are meaningless in a Python context (again the result of it    originally being ported from C) 4) Make it faster.
A test on the complete text of the King James Bible indicates that post-changes
the stemmer has entirely consistent results with the previous version and is
roughly 40% faster. It is also now threadsafe since nothing is stored in
any instance variables any more; instead things that used to be instance
variables are now passed between functions as arguments.