SemcorCorpusReader was returning incomplete sense information
semcor.py was creating lemma sense tags as strings of the form `<lemma>.<sensenum>`, which is incorrect: the sense number is actually POS-specific. With this fix, SemcorCorpusReader tries to retrieve the Lemma object directly from WordNet (using the sense key). If that fails, a string of the form `<lemma>.<pos>.<sensenum>` is given.