Call gs directly in nltk.tree._repr_png_, no shell
shell=True causes ghostscript to stick around with a prompt on
MacOS X

    GPL Ghostscript 9.06 (2012-08-08)
    Copyright (C) 2012 Artifex Software, Inc.  All rights reserved.
    This software comes with NO WARRANTY: see the file PUBLIC for details.
    GS>

Note that the shell=True was deliberately introduced in 0cc3d8a
to avoid a console window from popping up, possibly
related to Windows? If so, this patch may undo that change, but
maybe we can have the best of both worlds by setting the
STARTUPINFO.wShowWindow flag instead.

http://stackoverflow.com/questions/3172470/actual-meaning-of-shell-true-in-subprocess