Fixed "About" dialog issue with Python 3.x
Attribute name "_top" was misspelled, which resulted in an error when trying to show a fallback ShowText.
tkMessageBox have been renamed to messagebox in Python 3.x. I have added another import to take this into account. From my understanding, the fallback ShowText is not needed now, but I left it just in case.