Introduce extraneous training word to steal some probability mass for aligning to NULL token
In the original doctest, every target word was aligned to a non-NULL
source word. This posed a problem in imposing a probability distribution
P(target word|NULL), since this value is zero for all target words, yet
the sum of P(target word|NULL) over all target words has to be 1 in
order for it to be a proper probability distribution. As a result, some
probability values were grossly inflated so that P(target word|NULL)
will sum to 1.

To solve this, an extraneous word is introduced which legitimately
aligns to NULL. This diverts some probability mass to this word.

With real-world training data, this problem is less likely to occur, since
extraneous words are plentiful and the vocabulary sizes are larger.